Uploading UW application called Carol_FileIO
Carol_FileIO contains my first draft of the IO class. I also played around a little with the CreatePlaylist page just to get used to the Xaml controls, and uploading a picture.   Included in this push is a .txt file which is called Carol_UWP_notes.txt. This is the go to for any notes I you may need to use.