Updated Carols FileIO MusicApp with Player2
Updated Carols FileIO Music App with a second Media Player titled Player2. This player allows the user to select music to play using file picker, and allows user to select an image using file picker as well. The song and image are not stored but are played in real time. I chose to do this so that I had a better understanding of linking up files to the mediaplayer and I was able to successfully practice binding to the embedded mediaplayer. The GUI is still very simple until we decide upon the final implementation. ;-)