Adding UWP_MusicApp_Final and updating Carol_FileIO
Adding UWP_MusicApp_Final  which will be the final version of the music application. It is currently able to display music from MusicFolder, user can play music, and select music to be written to a playlist. User can name playlist and attach an image to the playlist only (not to each song). The playlist will be saved as a text file into the documents folder.
Remaining work in progress: Update the GUI, Update Mainpage display with more appropriate listview  with title header , play playlist music on mainpage (drop down button). We are getting there! ;-)