Updating UWP_MusicApp_Final
UWP_MusicApp_Final was updated to catch  the filenotfound exceptions if user selects an image outside of pictures top directory. It was changed in Playlist page as well as createplaylist page. This is a stopgap measure, as it prevents the application from crashing out when a file outside of the Pictures directory is specified.